Code Modularity Updated
